+++ src/services/http/middleware/RateLimiter.ts
@@ -188,7 +188,13 @@ class RateLimiter extends AbstractMiddleware {
     return res
       .status(429)
       .setHeader('Retry-After', String(consumeResult.retryAfter))
-      .json({ message: 'Too Many Requests' });
+      .json({
+        message: this.translate(
+          req,
+          'middleware.rateLimiter.tooManyRequests',
+          'Too Many Requests',
+        ),
+      });
   }
 }